+++ src/sll/include/sll/platform.h
@@ -65,8 +65,18 @@
  * \desc Docs!
  * \type const sll_char_t*
  */
-
 __SLL_EXTERNAL extern const sll_char_t* sll_platform_string;
+
+
+
+/**
+ * \flags var
+ * \name sll_environment
+ * \group platform
+ * \desc Docs!
+ * \type const sll_environment_t*
+ */
+__SLL_EXTERNAL extern const sll_environment_t* sll_environment;